target_thetaを0から2piにしてみた

+++ reRo_robotrace_board_code/Core/Inc/InvertedControl.hpp
@@ -80,7 +80,7 @@ public:
 	void start();
 	void stop();
 	void resetEstimatedTheta();
-	void setTargetOmega(float);
+	void setTargetOmega(double);
 
 	bool fallDown();
 	void getDytu(double *, double *);
